refactor auth server
